requirements.txt
asgiref==3.7.2
asttokens==2.0.8
backcall==0.2.0
colorama==0.4.5
debugpy==1.6.3
decorator==5.1.1
Django==5.0.3
django-ckeditor==6.7.1
django-ckeditor-5==0.2.15
django-cors-headers==4.3.1
django-js-asset==2.2.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
entrypoints==0.4
executing==1.1.0
ipykernel==6.16.0
ipython==8.5.0
jedi==0.18.1
joblib==1.4.2
jupyter-core==4.11.1
jupyter_client==7.3.5
matplotlib-inline==0.1.6
nest-asyncio==1.5.6
numpy==2.1.3
packaging==21.3
pandas==2.2.3
parso==0.8.3
pickleshare==0.7.5
pillow==10.4.0
prompt-toolkit==3.0.31
psutil==5.9.2
pure-eval==0.2.2
Pygments==2.13.0
PyJWT==2.9.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2024.2
pyzmq==26.2.0
scikit-learn==1.5.2
scipy==1.14.1
six==1.16.0
sqlparse==0.4.4
stack-data==0.5.1
threadpoolctl==3.5.0
tornado==6.2
traitlets==5.4.0
typing_extensions==4.10.0
tzdata==2024.1
wcwidth==0.2.5
whitenoise==6.8.2
